renamed Project from model specific IS1 to TobiiEyetracker, updated help patches
