+++ vvvv45/addonpack/lib/nodes/plugins/Browser (Devices TobiiEyetracker) help.v4p
@@ -1,5 +1,5 @@
 <!DOCTYPE PATCH  SYSTEM "http://vvvv.org/versions/vvvv45alpha29.5.dtd" >
-   <PATCH nodename="C:\Users\type name here\Documents\GitHub\vvvv-sdk\vvvv45\addonpack\lib\nodes\plugins\Browser (Devices IS1) help.v4p">
+   <PATCH nodename="C:\Users\type name here\Documents\GitHub\vvvv-sdk\vvvv45\addonpack\lib\nodes\plugins\Browser (Devices TobiiEyetracker) help.v4p">
    <BOUNDS type="Window" left="5925" top="2610" width="16305" height="10485">
    </BOUNDS>
    <NODE nodename="IOBox (String)" componentmode="InABox" id="3" systemname="IOBox (String)">
@@ -7,7 +7,7 @@
    </BOUNDS>
    <BOUNDS type="Box" left="150" top="150" width="5000" height="450">
    </BOUNDS>
-   <PIN pinname="Input String" slicecount="1" visible="0" values="|Browser (Devices IS1)|">
+   <PIN pinname="Input String" slicecount="1" visible="0" values="|Browser (Devices TobiiEyetracker)|">
    </PIN>
    <PIN pinname="Output String" visible="0">
    </PIN>
@@ -19,17 +19,9 @@
    </BOUNDS>
    <BOUNDS type="Box" left="150" top="550" width="5000" height="500">
    </BOUNDS>
-   <PIN pinname="Input String" slicecount="1" visible="0" values="|Eyetracker IS1 Node|">
+   <PIN pinname="Input String" slicecount="1" visible="0" values="|TobiiEyetracker Browser Node|">
    </PIN>
    <PIN pinname="Output String" visible="0">
-   </PIN>
-   </NODE>
-   <NODE nodename="Browser (Devices IS1)" componentmode="Hidden" id="0" systemname="Browser (Devices IS1)" filename="%VVVV%\addonpack\lib\nodes\plugins\IS1.dll">
-   <BOUNDS type="Node" left="1190" top="4405" width="9495" height="270">
-   </BOUNDS>
-   <BOUNDS type="Box" left="1190" top="4405" width="0" height="0">
-   </BOUNDS>
-   <PIN pinname="Update" slicecount="1" visible="1" values="0">
    </PIN>
    </NODE>
    <NODE systemname="IOBox (Value Advanced)" nodename="IOBox (Value Advanced)" componentmode="InABox" id="5">
@@ -50,8 +42,6 @@
    <PIN pinname="Y Input Value" slicecount="1" values="0">
    </PIN>
    </NODE>
-   <LINK srcnodeid="5" srcpinname="Y Output Value" dstnodeid="0" dstpinname="Update">
-   </LINK>
    <NODE nodename="IOBox (Value Advanced)" componentmode="InABox" id="6" systemname="IOBox (Value Advanced)">
    <BOUNDS type="Box" left="10620" top="3585" width="480" height="480">
    </BOUNDS>
@@ -72,8 +62,6 @@
    <PIN pinname="Descriptive Name" slicecount="1" values="Enable">
    </PIN>
    </NODE>
-   <LINK srcnodeid="6" srcpinname="Y Output Value" dstnodeid="0" dstpinname="Enable">
-   </LINK>
    <NODE nodename="IOBox (String)" componentmode="InABox" id="7" systemname="IOBox (String)">
    <BOUNDS type="Node" left="11130" top="3570" width="7170" height="270">
    </BOUNDS>
@@ -96,8 +84,6 @@
    <PIN pinname="Descriptive Name" slicecount="1" values="|Eyetracker Information|">
    </PIN>
    </NODE>
-   <LINK srcnodeid="0" srcpinname="Eyetracker Information" dstnodeid="8" dstpinname="Input Node">
-   </LINK>
    <NODE nodename="IOBox (String)" componentmode="InABox" id="9" systemname="IOBox (String)">
    <BOUNDS type="Node" left="2250" top="1110" width="6555" height="480">
    </BOUNDS>
@@ -156,8 +142,6 @@
    <PIN pinname="Descriptive Name" slicecount="1" values="Model">
    </PIN>
    </NODE>
-   <LINK srcnodeid="0" srcpinname="Model" dstnodeid="13" dstpinname="Input String">
-   </LINK>
    <NODE nodename="IOBox (String)" componentmode="InABox" id="15" systemname="IOBox (String)">
    <BOUNDS type="Box" left="3885" top="5220" width="795" height="240">
    </BOUNDS>
@@ -172,8 +156,6 @@
    <PIN pinname="Descriptive Name" slicecount="1" values="Name">
    </PIN>
    </NODE>
-   <LINK srcnodeid="0" srcpinname="Name" dstnodeid="15" dstpinname="Input String">
-   </LINK>
    <NODE nodename="IOBox (String)" componentmode="InABox" id="17" systemname="IOBox (String)">
    <BOUNDS type="Box" left="5235" top="5220" width="795" height="240">
    </BOUNDS>
@@ -188,8 +170,6 @@
    <PIN pinname="Descriptive Name" slicecount="1" values="|Product ID|">
    </PIN>
    </NODE>
-   <LINK srcnodeid="0" srcpinname="Product ID" dstnodeid="17" dstpinname="Input String">
-   </LINK>
    <NODE nodename="IOBox (String)" componentmode="InABox" id="18" systemname="IOBox (String)">
    <BOUNDS type="Box" left="9270" top="6150" width="795" height="240">
    </BOUNDS>
@@ -204,8 +184,6 @@
    <PIN pinname="Descriptive Name" slicecount="1" values="Status">
    </PIN>
    </NODE>
-   <LINK srcnodeid="0" srcpinname="Status" dstnodeid="18" dstpinname="Input String">
-   </LINK>
    <NODE nodename="IOBox (String)" componentmode="InABox" id="19" systemname="IOBox (String)">
    <BOUNDS type="Node" left="10125" top="6150" width="5190" height="1110">
    </BOUNDS>
@@ -234,8 +212,6 @@
    <PIN pinname="Descriptive Name" slicecount="1" values="Gerneration">
    </PIN>
    </NODE>
-   <LINK srcnodeid="0" srcpinname="Gerneration" dstnodeid="20" dstpinname="Input String">
-   </LINK>
    <NODE nodename="IOBox (String)" componentmode="InABox" id="21" systemname="IOBox (String)">
    <BOUNDS type="Box" left="7920" top="5220" width="795" height="240">
    </BOUNDS>
@@ -250,8 +226,6 @@
    <PIN pinname="Descriptive Name" slicecount="1" values="Version">
    </PIN>
    </NODE>
-   <LINK srcnodeid="0" srcpinname="Version" dstnodeid="21" dstpinname="Input String">
-   </LINK>
    <NODE nodename="IOBox (String)" componentmode="InABox" id="22" systemname="IOBox (String)">
    <BOUNDS type="Node" left="2535" top="5850" width="8805" height="270">
    </BOUNDS>
@@ -282,8 +256,6 @@
    <PIN pinname="Descriptive Name" slicecount="1" values="Initialized">
    </PIN>
    </NODE>
-   <LINK srcnodeid="0" srcpinname="Initialized" dstnodeid="23" dstpinname="Y Input Value">
-   </LINK>
    <NODE nodename="IOBox (String)" componentmode="InABox" id="24" systemname="IOBox (String)">
    <BOUNDS type="Node" left="11820" top="4935" width="4440" height="270">
    </BOUNDS>
@@ -298,4 +270,32 @@
    <PIN pinname="String Type" slicecount="1" values="MultiLine">
    </PIN>
    </NODE>
+   <NODE nodename="Browser (Devices TobiiEyetracker)" componentmode="Hidden" id="0" systemname="Browser (Devices TobiiEyetracker)" filename="%VVVV%\addonpack\lib\nodes\plugins\TobiiEyetracker.dll">
+   <BOUNDS type="Node" left="1190" top="4405" width="9495" height="270">
+   </BOUNDS>
+   <BOUNDS type="Box" left="1190" top="4405" width="0" height="0">
+   </BOUNDS>
+   <PIN pinname="Update" visible="1">
+   </PIN>
+   </NODE>
+   <LINK srcnodeid="5" srcpinname="Y Output Value" dstnodeid="0" dstpinname="Update">
+   </LINK>
+   <LINK srcnodeid="6" srcpinname="Y Output Value" dstnodeid="0" dstpinname="Enable">
+   </LINK>
+   <LINK srcnodeid="0" srcpinname="Eyetracker Information" dstnodeid="8" dstpinname="Input Node">
+   </LINK>
+   <LINK srcnodeid="0" srcpinname="Model" dstnodeid="13" dstpinname="Input String">
+   </LINK>
+   <LINK srcnodeid="0" srcpinname="Name" dstnodeid="15" dstpinname="Input String">
+   </LINK>
+   <LINK srcnodeid="0" srcpinname="Product ID" dstnodeid="17" dstpinname="Input String">
+   </LINK>
+   <LINK srcnodeid="0" srcpinname="Status" dstnodeid="18" dstpinname="Input String">
+   </LINK>
+   <LINK srcnodeid="0" srcpinname="Gerneration" dstnodeid="20" dstpinname="Input String">
+   </LINK>
+   <LINK srcnodeid="0" srcpinname="Version" dstnodeid="21" dstpinname="Input String">
+   </LINK>
+   <LINK srcnodeid="0" srcpinname="Initialized" dstnodeid="23" dstpinname="Y Input Value">
+   </LINK>
    </PATCH>
