+++ vvvv45/addonpack/src/Addonpack.csproj
@@ -225,9 +225,9 @@
       <Project>{AFEE382B-F971-4CF4-9F35-0112207097EB}</Project>
       <Name>ArduinoFirmata</Name>
     </ProjectReference>
-    <ProjectReference Include="nodes\plugins\Devices\IS1\IS1.csproj">
+    <ProjectReference Include="nodes\plugins\Devices\TobiiEyetracker\TobiiEyetracker.csproj">
       <Project>{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}</Project>
-      <Name>IS1</Name>
+      <Name>TobiiEyetracker</Name>
     </ProjectReference>
     <ProjectReference Include="nodes\plugins\Devices\LD2000\LD2000.csproj">
       <Project>{9DD0DF24-D149-44C1-A051-642F3608328C}</Project>
