+++ vvvv45/addonpack/src/Addonpack.sln
@@ -327,7 +327,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UtilsIL", "..\..\..\common\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShellExecute", "nodes\plugins\Windows\ShellExecute\ShellExecute.csproj", "{6421BE18-0B20-4448-90A9-B93908994B8A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IS1", "nodes\plugins\Devices\IS1\IS1.csproj", "{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TobiiEyetracker", "nodes\plugins\Devices\TobiiEyetracker\TobiiEyetracker.csproj", "{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
