+++ vvvv45/addonpack/src/nodes/plugins/Devices/TobiiEyetracker/CalibrationRunner.cs
@@ -10,7 +10,7 @@ using Tobii.Eyetracking.Sdk;
 
 #endregion
 
-namespace IS1
+namespace TobiiEyetracker
 {
     public class CalibrationRunner
     {
