+++ vvvv45/addonpack/src/nodes/plugins/Devices/TobiiEyetracker/Properties/AssemblyInfo.cs
