+++ vvvv45/addonpack/src/nodes/plugins/Devices/TobiiEyetracker/TobiiEyetracker.csproj
@@ -9,7 +9,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>VVVV.Nodes</RootNamespace>
-    <AssemblyName>IS1</AssemblyName>
+    <AssemblyName>TobiiEyetracker</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -74,7 +74,7 @@
     <Compile Include="EyeTrackerManager.cs" />
     <Compile Include="GazeDataDummySender.cs" />
     <Compile Include="HeadboxNode.cs" />
-    <Compile Include="IS1Node.cs" />
+    <Compile Include="EyetrackerNode.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TrackingNode.cs" />
     <Compile Include="XConfigNode.cs" />
