+++ vvvv45/addonpack/src/nodes/plugins/Devices/TobiiEyetracker/TobiiEyetracker.sln
@@ -1,18 +1,24 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 11.00
 # Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IS1", "IS1.csproj", "{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TobiiEyetracker", "TobiiEyetracker.csproj", "{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}.Debug|x86.ActiveCfg = Debug|x86
+		{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}.Debug|x86.Build.0 = Debug|x86
 		{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}.Release|x86.ActiveCfg = Release|x86
+		{7DA50EB5-CDBD-49C8-95CE-75A2C685F540}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
