+++ vvvv45/addonpack/src/nodes/plugins/Devices/TobiiEyetracker/TrackingNode.cs
@@ -18,10 +18,10 @@ using Tobii.Eyetracking.Sdk.Exceptions;
 #endregion usings
 
 
-namespace IS1
+namespace TobiiEyetracker
 {
     #region PluginInfo
-    [PluginInfo(Name = "Tracking", Category = "Devices", Version = "IS1", Help = "Eyetracker IS1 Node", Tags = "")]
+    [PluginInfo(Name = "Tracking", Category = "Devices", Version = "TobiiEyetracker", Help = "TobiiEyetracker Tracking Node", Tags = "", Author = "niggos, phlegma")]
     #endregion PluginInfo
